fix: 파일 추가 오류 수정 (#51)
