+++ packages/ui/src/icons/src/colored/IcDeleteCircle.tsx
@@ -1,0 +1,20 @@
+import * as React from 'react';
+import type { SVGProps } from 'react';
+const SvgIcDeleteCircle = (props: SVGProps<SVGSVGElement>) => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width={24}
+    height={24}
+    fill="none"
+    {...props}
+  >
+    <circle cx={12} cy={12} r={10} fill="#F2F3F6" />
+    <path
+      stroke="#999BB4"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      d="m8.668 8.666 6.667 6.667m0-6.667-6.667 6.667"
+    />
+  </svg>
+);
+export default SvgIcDeleteCircle;
