+++ packages/ui/src/icons/src/mono/index.ts
@@ -22,9 +22,9 @@ export { default as IcHeaderAlarm } from './IcHeaderAlarm';
 export { default as IcInfo } from './IcInfo';
 export { default as IcLink } from './IcLink';
 export { default as IcLinkCopy } from './IcLinkCopy';
+export { default as IcMail } from './IcMail';
 export { default as IcMailBtn } from './IcMailBtn';
 export { default as IcMessageBtn } from './IcMessageBtn';
-export { default as IcMail } from './IcMail';
 export { default as IcModify } from './IcModify';
 export { default as IcPlus } from './IcPlus';
 export { default as IcPlusCircle } from './IcPlusCircle';
